+++ db/migrate/20201211065754_create_bench_press_weight_records.rb
@@ -1,7 +1,7 @@
 class CreateBenchPressWeightRecords < ActiveRecord::Migration[6.0]
   def change
     create_table :bench_press_weight_records do |t|
-      t.float      :bench_press_weight
+      t.integer      :bench_press_weight
       t.date       :bench_press_day
       t.references :user,               foreign_key: true
       t.timestamps
